Save created files to S3
Once a file has been created it needs to be saved to S3. The knox
library is being used for this.

Configuration settings can be stored in a `config.json` file (example
provided) or read from environment settings.

Set Content-Type header

Due to the current naming convention, when a file is uploaded to S3,
the content type isn't being set properly. Until a file (better) naming
convention is put into place, the content type must be manually set.